Add files via upload
##Summary:

This script is a simple website scanner that checks whether a list of URLs or a single URL contains a specified string. If a URL is found to contain the string, the script prints a message saying "found" and the URL. If a URL does not contain the string, the script prints a message saying "fail" and the URL. The script also provides the option to save the results to an output file and to only show positive results.

##Usage:

This script is used from the command line and has the following options:

-uf/--url_file: A file containing a list of URLs to be checked.

-u/--single_url: A single URL to check.

-mc/--check_string: The string to check for in the response.

-o/--output: An output file name to save the results.

-p/--positive: Only show positive results (URLs that contain the check_string).

##Example:
$ cat dummy-site.txt | qsreplace 'fuzzme' | grep 'fuzzme'  | sort -u > test.txt

$ python script.py -uf test.txt -mc "fuzzme" -o output.txt  -p 

Note: The qsreplace tool is not part of this script and must be installed separately.

The script was developed by a developer named c0ded-human.